Revert "Update input.cc"
This reverts commit 19b2bfe251f97d2786c8713b38b2652450d77c69.

+++ input.cc
@@ -1744,18 +1744,18 @@ void read_parameterfile(int rank)
 #endif
     uint64_t pre_zseed{};
     if (zseed_input > 0) {
-      pre_zseed = zseed_input;  // random number seed
-      printout("using input.txt specified random number seed of %ld\n", pre_zseed);
+      pre_zseed = static_cast<uint64_t>(zseed_input);  // random number seed
+      printout("using input.txt specified random number seed of %lu\n", pre_zseed);
     } else {
       pre_zseed = std::random_device{}();
-      printout("randomly-generated random number seed is %ld\n", pre_zseed);
+      printout("randomly-generated random number seed is %lu\n", pre_zseed);
     }
 
     /// For MPI parallelisation, the random seed is changed based on the rank of the process
     /// For OpenMP parallelisation rng is a threadprivate variable and the seed changed according
     /// to the thread-ID tid.
     const uint64_t zseed = pre_zseed + static_cast<uint64_t>(13 * (rank * get_num_threads() + tid));
-    printout("rank %d: thread %d has zseed %ld\n", rank, tid, zseed);
+    printout("rank %d: thread %d has zseed %lu\n", rank, tid, zseed);
     rng_init(zseed);
     /// call it a few times
     for (int n = 0; n < 100; n++) {
